third_party/WebKit/LayoutTests/fast/events/drag-and-drop-autoscroll-inner-frame.html

<!DOCTYPE html>
<head>
<style type="text/css">
#scrollable {
    height: 200px;
    overflow: auto;
    border: solid 3px #cc0000;
    font-size: 80px;
}
</style>
</head>
<body>
For manual testing, drag and drop "Drop Me" downwards and then upwards.
<script src="../../resources/js-test.js"></script>
<script>

var x, y, middleTermScrollOffset;
var iframe, iframeDocument, draggable;

function setUpTest()
{
    if (!window.eventSender) {
        debug('Please run within DumpRenderTree');
        return;
    }

    iframe = document.getElementById('scrollable');
    // iframes start in readyState complete, but with the about:blank doc, make sure our doc is loaded.
    if (iframe.contentWindow.location.href == "about:blank" || iframe.contentDocument.readyState != "complete")
        iframe.onload = testIt;
    else
        testIt();
}

function testIt()
{
    eventSender.dragMode = false;

    iframe = document.getElementById('scrollable');
    iframeDocument = iframe.contentDocument;
    draggable = iframeDocument.getElementById('draggable');

    iframeDocument.addEventListener("scroll", recordScroll);

    // Grab draggable.
    x = iframe.offsetLeft + draggable.offsetLeft + 7;
    y = iframe.offsetTop + draggable.offsetTop + 7;

    eventSender.mouseMoveTo(x, y);
    eventSender.mouseDown();

    // Move mouse to the bottom autoscroll border belt.
    y = iframe.offsetTop + iframe.offsetHeight - 10;
    eventSender.mouseMoveTo(x, y);
}

function recordScroll(e)
{
    autoscrollTestPart1();
    iframeDocument.removeEventListener("scroll", recordScroll);
}

function recordScroll2(e)
{
    autoscrollTestPart2();
    iframeDocument.removeEventListener("scroll", recordScroll);
}

function autoscrollTestPart1()
{
    if (iframe.contentDocument.scrollingElement.scrollTop == 0) {
        testFailed("Autoscroll should have scrolled the iframe downwards, but did not");
        finishTest();
        return;
    }

    testPassed("Autoscroll should have scrolled the iframe downwards, and did.");

    middleTermScrollOffset = iframe.contentDocument.scrollingElement.scrollTop;
    iframeDocument.addEventListener("scroll", recordScroll2);

    // Move mouse to the upper autoscroll border belt.
    y = iframe.offsetTop + 10;
    eventSender.mouseMoveTo(x, y);
}

function autoscrollTestPart2()
{
    shouldBeTrue("iframe.contentDocument.scrollingElement.scrollTop < middleTermScrollOffset")
    finishTest();
}

function finishTest()
{
    eventSender.mouseUp();
    document.body.removeChild(iframe);
    finishJSTest();
}

description('Check autoscroll within an inner frame by drag-and-drop');
window.jsTestIsAsync = true;
window.onload = setUpTest;
</script>
<iframe id="scrollable" src="data:text/html,
<p id='draggable' draggable='true' style='cursor: hand;'>
    <b>Drag me!</b>
</p>
Try to drag and drop the text above in the input element at the bottom of this iframe. It should scroll. Then, try the way back.
<br><br>more<br>more<br>more<br>more<br>more<br>more<br>more<br>more<br>more<br>more<br>more<br>more<br><input>
"></iframe>
</body>
