third_party/WebKit/LayoutTests/fast/events/drag-and-drop-autoscroll-inner-frame-expected.txt

Check autoscroll within an inner frame by drag-and-drop

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Autoscroll should have scrolled the iframe downwards, and did.
PASS iframe.contentDocument.scrollingElement.scrollTop < middleTermScrollOffset is true
PASS successfullyParsed is true

TEST COMPLETE
For manual testing, drag and drop "Drop Me" downwards and then upwards.
